Hi fi/bc 657 chat (#130)
- Updated general styling to match the Hifi,  removed unused files,
functions
- added modals to edit the group chat name and errors/success toasts  
- fixed unread message badge 
- expand text field to max 6lines, added scrollbar  
- organized conversations' components and divided into different files
with those changes;
AC01, AC02, AC03, AC04, AC05 and AC08 [COMPLETED] 

AC06, AC07 and AC09 [NOT COMPLETED]
"Automated process of creating a group chat (team chat) and sending a
message by bootcampr admin when a team is created" [NOT COMPLETED]


BUG: If user receives a message while they are viewing the chat, it's
marked as unread and message badge appears. (This issue will be resolved
when we refactor the code)

[HiFi/bc-657-chat](https://bootcamper.atlassian.net/browse/BC-657)
