+++ src/components/ChatDialog/ChatDialogMain/ChatDialogMain.scss
@@ -1,37 +1,31 @@
 .chat-dialog-container {
+  display: flex;
+  flex-direction: column;
   background-color: #ffffff;
-  border-radius: 8px;
+  border-radius: 4px;
   box-shadow: 0 8px 17px -6px rgba(0, 0, 0, 0.7);
   cursor: default;
-  height: 520px;
-  padding: 0 20px;
-  position: fixed;
-  right: 154px;
+  height: 560px;
+  position: absolute;
+  right: 124px;
   top: 90px;
-  width: 295px;
+  width: 364px;
   z-index: 999999;
 
-  .chat-header-main,
-  .chat-header-messages {
-    align-items: center;
+  .chat-header {
     display: flex;
+    flex-direction: row;
+    align-items: center;
     justify-content: space-between;
-    margin: 0;
-    margin-left: -3px;
-    margin-right: -13px;
-
     .main-title {
-      font-size: 20px;
-      margin-left: 3px;
-      margin-top: 26px;
+      font-size: 24px;
+      margin: 0;
     }
 
     .back-arrow {
       align-items: center;
-      column-gap: 5px;
+      column-gap: 16px;
       display: flex;
-      margin-left: -10px;
-
       .group-link {
         cursor: pointer;
 
@@ -42,7 +36,10 @@
       }
 
       h5 {
-        font-size: 14px;
+        display: flex;
+        align-items: center;
+        column-gap: 8px;
+        font-size: 16px;
         font-weight: 400;
         margin: 11px 0;
         max-width: 225px;
@@ -50,42 +47,29 @@
         text-overflow: ellipsis;
         white-space: nowrap;
       }
-
-      &.messages {
-        margin-bottom: -3px;
-        margin-top: -3px;
-      }
     }
-
     .main-icons {
       align-items: center;
-      column-gap: 8px;
+      column-gap: 24px;
       display: flex;
     }
-  }
+    &.main {
+      padding: 16px;
+    }
+    &.room {
+      padding: 8px 16px;
 
-  .main-icons.home {
-    margin-right: 13px;
-  }
-
-  .chat-header-messages {
-    border-bottom: 1px solid #86888a;
-    margin: 0 -20px;
-    padding: 8px 7px 8px 17px;
-
-    .back-arrow {
-      h5 {
-        margin: 0;
-      }
+      border-bottom: 1px solid #86888a;
     }
   }
-
+  .chat-body {
+    display: flex;
+    flex-direction: column;
+    flex: 1;
+    padding: 0 8px 16px 8px;
+  }
   svg {
     cursor: pointer;
-
-    &:hover {
-      color: #f29e34;
-    }
   }
 
   p {
