+++ src/components/ChatDialog/Conversations/Conversations.scss
@@ -2,49 +2,12 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-
-  .search {
-    align-items: center;
-    display: flex;
-    justify-content: center;
-    margin: 4px 0;
-
-    input {
-      background-color: #ecebeb;
-      border: none;
-      border-radius: 4px;
-      height: 35px;
-      padding-left: 40px;
-      width: 100%;
-    }
-
-    ::placeholder {
-      color: #313335;
-      font-size: 12px;
-    }
-
-    svg {
-      cursor: default;
-      left: 28px;
-      position: absolute;
-      top: 76px;
-
-      &:hover {
-        color: black;
-      }
-    }
-  }
-
   .conversations-list {
     align-items: center;
-    display: flex;
-    flex-direction: column;
-    margin: 15px -18px 0 -15px;
     max-height: 383px;
     overflow-x: hidden;
     overflow-y: auto;
-    padding: 0 8px 0 15px;
-
+    width: 100%;
     &::-webkit-scrollbar {
       background-color: #f5f5f500;
       height: 8px;
@@ -62,14 +25,12 @@
       opacity: 1;
     }
   }
-
   .conversation-grid {
     align-items: center;
-    column-gap: 10px;
+    column-gap: 16px;
     cursor: pointer;
     display: flex;
-    padding: 8px 0;
-    width: 100%;
+    padding: 8px;
 
     .unread-dot-visible {
       color: #fa9413;
@@ -85,7 +46,7 @@
     }
 
     .thumbnail-right {
-      align-content: center;
+      column-gap: 8px;
       display: flex;
       flex-grow: 1;
       justify-content: space-between;
@@ -93,6 +54,7 @@
       .dot-grid {
         align-items: center;
         display: flex;
+        margin-right: -8px;
 
         p {
           align-self: flex-end;
@@ -105,25 +67,26 @@
     .thread-details-grid {
       align-items: flex-start;
       display: flex;
+      flex-grow: 1;
       flex-direction: column;
-      row-gap: 3px;
+      row-gap: 4px;
 
       .unread-message {
         font-weight: 600;
       }
 
       h5 {
-        font-size: 13px;
+        font-size: 16px;
         font-weight: 400;
+        line-height: 24px;
         margin: 0;
-        max-width: 195px;
         overflow: hidden;
         text-overflow: ellipsis;
         white-space: nowrap;
       }
 
       p {
-        font-size: 11px;
+        font-size: 14px;
         margin: 0;
         max-width: 190px;
         overflow: hidden;
