+++ src/components/ChatDialog/MemberThumbnail/MemberThumbnail.scss
@@ -15,11 +15,14 @@
     }
 
     h5 {
-      font-size: 13px;
+      font-size: 16px;
+      line-height: 20px;
     }
 
     p {
-      font-size: 11.5px;
+      font-size: 14px;
+      line-height: 20px;
+      color: #616161;
     }
   }
 }
