+++ src/components/ChatDialog/NewChatRoom/NewChatRoom.scss
@@ -1,37 +1,33 @@
-.new-chat-room-container {
+.new-chat-container {
   display: flex;
   flex-direction: column;
-  justify-content: space-evenly;
-
+  flex-grow: 1;
+  justify-content: space-between;
   .checkbox {
     background-color: white;
-    border: 1px solid #022888;
-    border-radius: 4px;
+    border: 2px solid #212121;
+    border-radius: 2px;
     cursor: pointer;
-    font-size: 25px;
     height: 16px;
-    margin: 0 1px;
     width: 16px;
   }
 
   .checkbox-checked,
   .all-checked {
     align-items: center;
-    background-color: #022888;
-    border: 1px solid #022888;
-    border-radius: 4px;
+    background-color: #1a237e;
+    border: 2px solid #1a237e;
+    border-radius: 2px;
     color: white;
     cursor: pointer;
     display: flex;
-    font-size: 25px;
     height: 16px;
     justify-content: center;
-    margin: 0 1px;
     width: 16px;
 
     .checkmark {
-      font-size: 14px;
-      stroke-width: 0.5px;
+      font-size: 10px;
+      stroke-width: 1.75px;
 
       &:hover {
         color: white;
@@ -39,120 +35,59 @@
     }
   }
 
-  section {
-    height: 420px;
-    margin: -10px -20px 0 -20px;
-    overflow-x: hidden;
-    overflow-y: auto;
-    padding: 0 8px 0 8px;
-
-    &::-webkit-scrollbar {
-      background-color: #f5f5f500;
-      height: 8px;
-      width: 3px;
-    }
-
-    &::-webkit-scrollbar-thumb {
-      background-color: rgba(0, 0, 0, 0.2);
-      border-radius: 4px;
-      opacity: 0;
-      transition: opacity 0.2s ease-in-out;
-    }
-
-    &:hover::-webkit-scrollbar-thumb {
-      opacity: 1;
-    }
-
-    .select-all-grid {
-      align-items: center;
-      column-gap: 8px;
-      display: flex;
-    }
-
-    label {
-      font-size: 12px;
-    }
-
-    .to-container {
-      align-items: center;
-      background-color: white;
-      border-bottom: 1px solid #86888a;
-      column-gap: 8px;
-      display: flex;
-      flex-shrink: 0;
-      height: fit-content;
-      margin: 0 -8px 8px -8px;
-      min-height: 35px;
-      padding: 0 8px;
-      padding-top: 10px;
-      position: sticky;
-      top: 0;
-
-      p {
-        margin: 0;
-      }
-
-      .to-member-grid {
-        column-gap: 8px;
-        display: flex;
-        flex-wrap: wrap;
-        max-height: 58px;
-        overflow-y: auto;
-        padding: 5px 0;
-        row-gap: 6px;
-
-        .member-bar {
-          align-items: center;
-          background-color: #dff5ff;
-          border-radius: 4px;
-          column-gap: 3px;
-          display: flex;
-          padding: 3px 4px;
-        }
-      }
-    }
-
-    p {
-      font-size: 12px;
-      font-weight: 400;
-      margin: 0;
-      margin-bottom: 8px;
-      text-align: left;
-    }
-
-    h3 {
-      font-size: 16px;
-      text-align: left;
-    }
-
-    .members-container {
-      display: flex;
-      flex-direction: column;
-      margin: 8px -3px 8px 0;
-      padding: 0 3px 0 0;
-      row-gap: 2px;
-
-      .member-grid,
-      .member-selected {
-        align-items: center;
-        column-gap: 8px;
-        cursor: pointer;
-        display: flex;
-        padding: 8px 8px 8px 0;
-
-        &:hover {
-          background-color: #fcebd6;
-          border-radius: 8px;
-        }
-      }
-
-      .member-selected {
-        background-color: #fcebd6;
-        border-radius: 8px;
-      }
-    }
+  &::-webkit-scrollbar {
+    background-color: #f5f5f500;
+    height: 8px;
+    width: 3px;
   }
 
+  &::-webkit-scrollbar-thumb {
+    background-color: rgba(0, 0, 0, 0.2);
+    border-radius: 4px;
+    opacity: 0;
+    transition: opacity 0.2s ease-in-out;
+  }
+
+  &:hover::-webkit-scrollbar-thumb {
+    opacity: 1;
+  }
+
+  .members-invite {
+    font-size: 16px;
+    line-height: 24px;
+    font-family: 'Roboto';
+    display: flex;
+    flex-direction: column;
+    row-gap: 32px;
+    padding: 8px 8px 8px 8px;
+    p {
+      margin: 0;
+      padding: 0;
+      text-align: left;
+    }
+
+    .members-invite-select {
+      display: flex;
+      flex-direction: column;
+      row-gap: 16px;
+      overflow-y: auto;
+
+      .members-grid {
+        display: flex;
+        flex-direction: column;
+        row-gap: 8px;
+        .member-grid {
+          display: flex;
+          align-items: center;
+          column-gap: 16px;
+        }
+      }
+    }
+    .members-invite-info {
+      font-size: 14px;
+      line-height: 20px;
+    }
+  }
   .no-remaining-members {
     align-items: center;
     display: flex;
@@ -172,22 +107,15 @@
     }
   }
 
-  button {
+  .chat-button {
     background-color: #fa9413;
     border: none;
     border-radius: 4px;
     cursor: pointer;
-    font-size: 15px;
-    height: 42px;
-    margin: 5px -5px 0px -5px;
-
-    &:hover {
-      background-color: #f7a53b;
-    }
-
+    margin: 8px 8px 0 8px;
+    text-transform: none;
     &:disabled {
-      background-color: #ecebeb;
-      cursor: default;
+      opacity: 0.5;
     }
   }
 }
