+++ src/components/CommonModal/CommonModal.scss
@@ -53,7 +53,7 @@
 
   &.input:focus-visible {
     border: none;
-    outline: 2px solid #1976d2;
+    outline: 2px solid #0d47a1;
     outline-offset: -1px;
   }
